Fix regression

+++ src/main/java/delta/games/lotro/tools/extraction/items/MainDatItemsLoader.java
@@ -529,7 +529,7 @@ public class MainDatItemsLoader
         LOGGER.warn("Legendary item with no main legacy (DPS or not): {}",item);
       }
       @SuppressWarnings("unused")
-      int combatPropertyType=((Integer)properties.getProperty("Item_RequiredCombatPropertyType")).intValue();
+      Integer combatPropertyType=(Integer)properties.getProperty("Item_RequiredCombatPropertyType");
       // Seems that each legendary item (4764 occurrences) has either ItemAdvancement_CombatDPSLevel or ItemAdvancement_CombatPropertyModLevel
       // depending on whether it has a DPS main legacy of a non-DPS main legacy
       Integer combatDpsLevel=(Integer)properties.getProperty("ItemAdvancement_CombatDPSLevel");
